feat: move exception package from user to model
