+++ domain/usecase/src/test/java/co/irond/crediya/usecase/user/UserUseCaseTest.java
@@ -1,7 +1,7 @@
 package co.irond.crediya.usecase.user;
 
 import co.irond.crediya.model.user.User;
-import co.irond.crediya.model.user.exceptions.CrediYaException;
+import co.irond.crediya.model.exceptions.CrediYaException;
 import co.irond.crediya.model.user.gateways.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
